3주차 강의 실습.
아이템 시스템과 게임 흐름 로직 구현.

ItemInterfaceClass.

GameStateClass.

BaseItemClass.

BigCoinItemClass.

SmallCoinItemClass.

MineItemClass.

HealingItemClass.

+++ Source/PlayerCharacter/Public/HealingItem.h
@@ -13,7 +13,7 @@ public:
 	AHealingItem();
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Item")
-	float HealAmount;
+	int32 HealAmount;
 
 	virtual void ActivateItem(AActor* Activator) override;
